On branch feature/setup  Your branch is up-to-date with 'origin/feature/setup'.
	modified:   src/alinea/phenomenal/mesh/algorithms.py
	modified:   src/alinea/phenomenal/mesh/vtk_transformation.py
	modified:   src/alinea/phenomenal/multi_view_reconstruction/multi_view_reconstruction.py
	modified:   src/alinea/phenomenal/multi_view_reconstruction/multi_view_reconstruction_octree.py
	modified:   src/alinea/phenomenal/segmentation_2d/skeletonize.py
	renamed:    test/test_mesh/test_mesh.py -> test/test_mesh/test_algorithms.py
	modified:   test/test_multi_view_reconstruction/test_multi_view_reconstruction.py
	new file:   test/test_segmentation_2D/test_skeletonize.py
	renamed:    test/test_segmentation/test_create_graph.py -> test/test_segmentation_3D/test_create_graph.py
	renamed:    test/test_segmentation/test_graph_connect.py -> test/test_segmentation_3D/test_graph_connect.py
	new file:   test/test_segmentation_3D/test_maize.py
	renamed:    test/test_segmentation/test_skeleton.py -> test/test_segmentation_3D/test_skeleton.py

+++ src/alinea/phenomenal/segmentation_2d/skeletonize.py
@@ -30,7 +30,7 @@ def skeletonize_thinning(image):
     :param image: binary image with 0 or 255
     :return: skeleton of a binary image.
     """
-    img = image.copy()
+    img = image.copy().astype(numpy.uint8)
 
     img[img == 255] = 1
     skeleton = skimage.morphology.skeletonize(img)
@@ -47,7 +47,7 @@ def skeletonize_erode_dilate(image):
     :param image: binary image with 0 or 255
     :return: skeleton of a binary image.
     """
-    img = image.copy()
+    img = image.copy().astype(numpy.uint8)
 
     skeleton = numpy.zeros(img.shape, numpy.uint8)
